+++ nengo_bones/tests/test_pr_number.py
@@ -43,7 +43,7 @@ def test_pr_number(monkeypatch, tmp_path, mode):
             project_name: NengoBones
             pkg_name: nengo-bones
             repo_name: {repo}
-        """,
+            """,
         )
 
     if mode == "repo":