+++ CHANGES.rst
@@ -68,6 +68,9 @@ Release History
 - Increased minimum ``black`` version to 21.12b0. (`#104`_, `#159`_)
 - License fields in ``setup.py`` are now set using ``license_rst.type``. (`#161`_)
 - Trove classifiers in ``setup.py`` are now sorted alphabetically. (`#161`_)
+- Updated Nengo license and renamed to ABR license. (`#157`_)
+- The license type is now specified through a ``license`` key at the top level.
+  If not specified, it defaults to ``proprietary``. (`#157`_)
 
 **Fixed**
 
@@ -98,6 +101,7 @@ Release History
 .. _#145: https://github.com/nengo/nengo-bones/pull/145
 .. _#150: https://github.com/nengo/nengo-bones/pull/150
 .. _#151: https://github.com/nengo/nengo-bones/pull/151
+.. _#157: https://github.com/nengo/nengo-bones/pull/157
 .. _#159: https://github.com/nengo/nengo-bones/pull/159
 .. _#160: https://github.com/nengo/nengo-bones/pull/160
 .. _#161: https://github.com/nengo/nengo-bones/pull/161